refactored magicStart to allow for easier furniture additions.
The original implementation had furniture creation integrate tightly
into the rendering of the Window. This commit pulls furniture out into
its own class, and provides a `draw()` and `create()` function for each
furniture.

This means that the main gui window can simply call <object>.draw() or
<object>.create() and everything for that specific furniture will be
handled. This includes all rendering of icons and text boxes for special
calculations.

Adding a new furniture object is trivial, in that, you need only copy
the existing example class (or us inheritance), then modify the areas
specific to that furnitures needs. Lastly, add it to the dictionary at
the top of the file and it will automatically show up in the drop down.

Being that furniture can vary greatly, it doesn't make a lot of sense to
try to stuff them all into similar functions. So keep them separate, but
organized.